requirements.txt
torch
transformers
pillow
numpy
matplotlib
tqdm